Move firebase class to a separate utils folder
