+++ App.js
@@ -9,8 +9,8 @@ import { Ionicons } from '@expo/vector-icons';
 import firebase from 'firebase';
 
 import { getImage } from './constants/ImageFilepaths';
-import { initializeFirebase } from './firebase/firebase';
 import AppNavigator from './navigation/AppNavigator';
+import { initializeFirebase } from './util/firebase/firebase';
 
 export default function App(props) {
   const [isLoadingComplete, setLoadingComplete] = useState(false);
