+++ util/firebase/firebase.js
@@ -1,6 +1,6 @@
 import * as firebase from 'firebase';
 
-import getEnvVars from '../environment';
+import getEnvVars from '../../environment';
 const { firebaseAuthDomain, firebaseDatabaseURL, firebaseStorageBucket } = getEnvVars();
 
 export async function initializeFirebase() {
